+++ packages/ui/src/components/va-select/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.9.0
+- Now in `content` slot `value` is not array in single mode, but array in multiple mode. Use `valueArray` if you need array value in single/multiple modes.
 - Allow overriding `keepAnchorWidth`, `cloneOnContentClick`, `triggers`, `offset` props
 - Fix incorrect search value if `model-value` is null or undefined
 - Add slot types